Add muscle validation to exercise form
