+++ components/muscle-combobox.tsx
@@ -8,28 +8,7 @@ import { UseFormSetValue } from "react-hook-form";
 import { z } from "zod";
 import { exerciseFormSchema } from "./exercise-dialog-form";
 import { ScrollArea } from "./ui/scroll-area";
-
-const muscles = [
-  { value: "chest", label: "Peitoral" },
-  { value: "back", label: "Costas" },
-  { value: "shoulders", label: "Ombro" },
-  { value: "biceps", label: "Bíceps" },
-  { value: "triceps", label: "Tríceps" },
-  { value: "forearms", label: "Antebraço" },
-  { value: "abs", label: "Abdômen" },
-  { value: "quadriceps", label: "Quadríceps" },
-  { value: "hamstrings", label: "Isquiotibiais" },
-  { value: "glutes", label: "Glúteos" },
-  { value: "adductors", label: "Adutores" },
-  { value: "abductors", label: "Abdutores" },
-  { value: "calves", label: "Panturrilha" },
-  { value: "fullBody", label: "Corpo inteiro" },
-  { value: "upperBody", label: "Corpo superior" },
-  { value: "lowerBody", label: "Corpo inferior" },
-  { value: "cardio", label: "Cardio" },
-  { value: "stretching", label: "Alongamento" },
-  { value: "other", label: "Outro" },
-] as const;
+import { muscles } from "@/lib/muscles";
 
 type MuscleComboboxProps = {
   value: string;
