+++ lib/muscles.ts
@@ -1,0 +1,23 @@
+const muscles = [
+  { value: "chest", label: "Peitoral" },
+  { value: "back", label: "Costas" },
+  { value: "shoulders", label: "Ombro" },
+  { value: "biceps", label: "Bíceps" },
+  { value: "triceps", label: "Tríceps" },
+  { value: "forearms", label: "Antebraço" },
+  { value: "abs", label: "Abdômen" },
+  { value: "quadriceps", label: "Quadríceps" },
+  { value: "hamstrings", label: "Isquiotibiais" },
+  { value: "glutes", label: "Glúteos" },
+  { value: "adductors", label: "Adutores" },
+  { value: "abductors", label: "Abdutores" },
+  { value: "calves", label: "Panturrilha" },
+  { value: "fullBody", label: "Corpo inteiro" },
+  { value: "upperBody", label: "Corpo superior" },
+  { value: "lowerBody", label: "Corpo inferior" },
+  { value: "cardio", label: "Cardio" },
+  { value: "stretching", label: "Alongamento" },
+  { value: "other", label: "Outro" },
+] as const;
+
+export { muscles };
